+++ test/mocks/newrelic_rpm.rb
@@ -10,5 +10,17 @@ module NewRelic
 
     def self.increment_metric(name, count = 1)
     end
+
+    def self.linking_metadata
+      {"entity.name"=>"Entity Name", "entity.type"=>"SERVICE", "hostname"=>"hostname"}
+    end
+
+    module Tracer
+      def self.current_trace_id
+      end
+
+      def self.current_span_id
+      end
+    end
   end
 end