+++ test/appender/http_test.rb
@@ -82,12 +82,12 @@ module Appender
       end
 
       it "uses the ENV proxy if specified" do
-        old_env_proxy = ENV["http_proxy"]
+        old_env_proxy = ENV.fetch("http_proxy", nil)
         ENV["http_proxy"] = "http://user:password@proxy.example.com:12345"
         Net::HTTP.stub_any_instance(:start, true) do
           appender = SemanticLogger::Appender::Http.new(url: "http://ruby-lang.org:8088/path")
 
-          proxy_uri = URI.parse(ENV["http_proxy"])
+          proxy_uri = URI.parse(ENV.fetch("http_proxy", nil))
           assert(appender.http.proxy?)
           assert(appender.http.proxy_from_env?)
           assert_equal(proxy_uri.host, appender.http.proxy_address)
@@ -100,7 +100,7 @@ module Appender
       end
 
       it "doesn't use the ENV proxy if explicity requested" do
-        old_env_proxy = ENV["http_proxy"]
+        old_env_proxy = ENV.fetch("http_proxy", nil)
         ENV["http_proxy"] = "http://user:password@proxy.example.com:12345"
         Net::HTTP.stub_any_instance(:start, true) do
           appender = SemanticLogger::Appender::Http.new(url: "http://ruby-lang.org:8088/path", proxy_url: nil)