+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
+- Move `newrelic_rpm.rb` mock to the `test/mocks` directory
+- Add `add_mocks_to_load_path`
 
 ## [4.13.0]