+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Add payload logging to `payload_includes` Minitest helper functionality.
+- Fix bug in `payload_includes` testing functionality where if a key is not in the payload the test would pass.
+
 ## [4.13.0]
 
 - Add Minitest helper methods to assist with asserting logging events.