+++ CHANGELOG.md
@@ -1,14 +1,32 @@
 # Change Log
+
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
-- The Logfmt formatter can now parse Hashes and Arrays correctly.
-- Fixes a race condition in `SemancicLogger.reopen`.
+- Move `newrelic_rpm.rb` mock to the `test/mocks` directory
+- Add `add_mocks_to_load_path`
+- Add NewRelicLogs formatter & appender
 - Move initialization of `events` array in `CaptureLogEvents` to constructor.
 - Add `clear` method to `CaptureLogEvents`.
 
-### Changed
+- Add payload logging to `payload_includes` Minitest helper functionality.
+- Fix bug in `payload_includes` testing functionality where if a key is not in the payload the test would pass.
+
+- Quote 3.0 in CI configuration to prevent truncation.
+
+## [4.13.0]
+
+- Add Minitest helper methods to assist with asserting logging events.
+
+## [4.12.0]
+
+- Add SyslogCee formatter for Syslog.
+- The Logfmt formatter can now parse Hashes and Arrays correctly.
+- Fixes a race condition in `SemanticLogger.reopen`.
+- Ability to directly set proxy for `SemanticLogger::Appender::Http`
+- Fixed log level signal handler to include fatal and error levels.
+- Fixes #218 Calls Sentry.init inside sentry_ruby appender only if it is not initialized
 - Contributor experience related to RuboCop was improved with the
   following changes:
     - New .rubocop\_todo.yml to hold RuboCop offenses and enable a clean
@@ -18,9 +36,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
     - Updated RuboCop's target version to Ruby 2.7.5.
     - Updated minimum Ruby version to 2.7.5 as earlier versions are
       end-of-life.
-- New Feature: Ability to directly set proxy for `SemanticLogger::Appender::Http`
-- Fixed log level signal handler to include fatal and error levels.
-- Fixes #218 Calls Sentry.init inside sentry_ruby appender only if it is not initialized
+- Add mutexes to `SemanticLogger.sync!` in case some users are still using it in a 
+  multi-threaded environment.
 
 ## [4.11.0]